+++ wf_parser/src/lib.rs
@@ -5,15 +5,3 @@ mod message_code_parser;
 mod wf_header;
 
 pub use message_code_parser::MessageCodeParser;
-
-pub trait MessageHeader {
-    type Target: ?Sized;
-
-    fn prefix(&self) -> &Self::Target;
-    fn version(&self) -> &Self::Target;
-    fn encryption_indicator(&self) -> &Self::Target;
-    fn duress_indicator(&self) -> &Self::Target;
-    fn message_code(&self) -> &Self::Target;
-    fn reference_indicator(&self) -> &Self::Target;
-    fn referenced_message(&self) -> &Self::Target;
-}
